Refactor : 코드정리2
