+++ src/main/java/com/sdy/bbbb/repository/ImageRepository.java
@@ -1,7 +1,7 @@
 package com.sdy.bbbb.repository;
 
 import com.sdy.bbbb.entity.Image;
-import com.sdy.bbbb.querydsl.ImageRepositoryCustom;
+import com.sdy.bbbb.repository.querydsl.ImageRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ImageRepository extends JpaRepository<Image, Long>, ImageRepositoryCustom {
