+++ src/main/java/com/sdy/bbbb/repository/querydsl/ImageRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.sdy.bbbb.querydsl;
+package com.sdy.bbbb.repository.querydsl;
 
 public interface ImageRepositoryCustom {
 }
