--- src/main/java/com/sdy/bbbb/sentry/ExampleService.java
@@ -1,27 +1,0 @@
-package com.sdy.bbbb.sentry;
-
-import io.sentry.Sentry;
-import io.sentry.SentryEvent;
-import io.sentry.SentryLevel;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-import java.util.UUID;
-
-@Slf4j
-@Service
-public class ExampleService {
-    public void example1() {
-        log.info("Hello World1");
-    }
-
-
-    public void example3() {
-        try {
-            throw new Exception("This is a test.");
-        } catch (Exception e) {
-            Sentry.captureException(e);
-        }
-    }
-}
-
