+++ src/main/java/com/sdy/bbbb/service/MyPageService.java
@@ -5,7 +5,7 @@ import com.sdy.bbbb.dto.response.*;
 import com.sdy.bbbb.entity.*;
 import com.sdy.bbbb.exception.CustomException;
 import com.sdy.bbbb.exception.ErrorCode;
-import com.sdy.bbbb.querydsl.CommentRepositoryImpl;
+import com.sdy.bbbb.repository.querydsl.CommentRepositoryImpl;
 import com.sdy.bbbb.repository.*;
 import com.sdy.bbbb.s3.S3Uploader;
 import com.sdy.bbbb.util.ServiceUtil;
